deployments/marketplace-management/requirements.txt
databricks-sdk==0.74.0
termcolor==3.2.0
yamale==6.1.0
